unsigbiginteger

+++ database/migrations/2023_03_27_044600_create_study_classes_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -17,8 +16,8 @@ return new class extends Migration
             $table->increments('id');
             $table->string('slug')->nullable();
             $table->string('name');
-            $table->unsignedInteger('id_major');
-            $table->unsignedInteger('id_level');
+            $table->unsignedBigInteger('id_major');
+            $table->unsignedBigInteger('id_level');
             $table->tinyInteger('status')->default(1);
             $table->timestamps();
             $table->softDeletes();
